testing fails on TEST_CASE('All filtering bits used')  when using uint32_t for LSHDatatype, but not for uint64_t

+++ include/puffinn/typedefs.hpp
@@ -19,7 +19,7 @@ namespace puffinn {
 
 
     //This defines the current LshDatatype to be used -- would prefer to give this information as parameter to puffinn
-    using LshDatatype = HammingType<uint32_t>;
+    using LshDatatype = HammingType<uint64_t>;
 
     std::default_random_engine generator(std::chrono::system_clock::now().time_since_epoch().count());
